Organize package structure and refactor classes in the cart package
