+++ src/main/java/com/ecommerce/api/cart/service/CartRepository.java
@@ -1,5 +1,6 @@
-package com.ecommerce.api.cart;
+package com.ecommerce.api.cart.service;
 
+import com.ecommerce.api.cart.model.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
