+++ src/main/java/com/ecommerce/api/cart/service/CartService.java
@@ -1,5 +1,8 @@
-package com.ecommerce.api.cart;
+package com.ecommerce.api.cart.service;
 
+import com.ecommerce.api.cart.model.Cart;
+import com.ecommerce.api.cart.response.CartResponse;
+import com.ecommerce.api.cart.service.CartRepository;
 import com.ecommerce.api.cart_item.CartItemService;
 import com.ecommerce.api.cart_item.RequestCartItemDto;
 import com.ecommerce.api.user.User;
