+++ src/main/java/com/ecommerce/api/order/OrderService.java
@@ -5,8 +5,8 @@ import com.ecommerce.api.product.Product;
 import com.ecommerce.api.product.ProductRepository;
 import com.ecommerce.api.utility.EnumConstants;
 import com.ecommerce.api.utility.NotFoundException;
-import com.ecommerce.api.cart.Cart;
-import com.ecommerce.api.cart.CartRepository;
+import com.ecommerce.api.cart.model.Cart;
+import com.ecommerce.api.cart.service.CartRepository;
 import com.ecommerce.api.order_item.RequestOrderItemDto;
 import com.ecommerce.api.order_timeline.OrderTimeline;
 import org.modelmapper.ModelMapper;
