+++ src/main/java/com/ecommerce/api/user/User.java
@@ -2,7 +2,7 @@ package com.ecommerce.api.user;
 
 import com.ecommerce.api.order.Order;
 import com.ecommerce.api.utility.EnumConstants;
-import com.ecommerce.api.cart.Cart;
+import com.ecommerce.api.cart.model.Cart;
 import jakarta.persistence.*;
 import lombok.*;
